Remove AdvancedUtils and move its methods to AIUtils
